refactor: move demo CLI commands into demo/cli/ package
