+++ nbkp/demo/cli/__init__.py
@@ -1,4 +1,4 @@
-"""Demo CLI app definition."""
+"""Demo CLI sub-app."""
 
 import typer
 from rich.console import Console
@@ -10,3 +10,6 @@ app = typer.Typer(
 )
 
 console = Console()
+
+from . import output_cmd as _output_cmd  # noqa: E402, F401
+from . import seed_cmd as _seed_cmd  # noqa: E402, F401
